Added support for `[u8]` inputs in `Location`
This fixes a problem where using `final_parser()` on a byte string
(`[u8]`) input would fail to compile because `Location` did not know how
to handle anything other than `str`.

I tried to keep this as small and simple as possible. An alternative fix
would have been to adjust `Location::locate_tail()` to take generic
parameters that could handle both `[u8]` and `str`, but that would be
significantly more complicated and harder to understand.